connect payment api

+++ microservices.springframework/user-package-service/src/main/java/swd392/userpackageservice/application/dto/UserPackageResponse.java
@@ -1,8 +1,8 @@
 package swd392.userpackageservice.application.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -16,6 +16,6 @@ public class UserPackageResponse {
     private UUID id;
     private UUID userId;
     private UUID packageId;
-    private float price;
+    private BigDecimal price;
     private Instant transactionDate;
 }
